Make rocket runnable on cargo-web
With these few changes, rocket can be run on using `cargo +nightly web
start --target-webasm --release`:

* Crate changed from cdylib to regular binary (which happens to have an
  empty main because JS drives it).
* Paths changed in index.html and post_build.py to where cargo-web
  expects and serves them (html → static, program.wasm → rocket.wasm).

The post_build.py script was left in place because cargo-web does not
run the GC itself (though it probably should).

The program can now either be built as before, equivalently with `web
build --target-webasm --release`, or served with the above `web start`
command.

Note that the wasm32-unknown-unknown mode of cargo-web is still very
immature, (as of today) not even released (outside of its master
branch), and the only benefit it gives so far is running a web server
and auto-compiling whenever the code changes (so you only need to
refresh in the browser).